Style changes / Split signUp - Split signUp form into a component;
