+++ src/interfaces/IErrorMessages.ts
@@ -1,0 +1,9 @@
+export enum signUpCnpjErrorMessages {
+    EMPTY = "Por favor, digite um CNPJ",
+    INVALID = "Por favor, digite um CNPJ válido"
+}
+
+export enum signUpPhoneErrorMessages {
+    EMPTY = "Por favor, digite um telefone",
+    INVALID = "Por favor, digite um telefone válido"
+}
