+++ src/screens/SignUp/SignUpForm/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './SignUpForm'
